autowire/requirements.txt
balena-sdk==9.5.0
PluginBase